+++ scripts/_helpers.py
@@ -11,10 +11,10 @@ import os
 import re
 import time
 from bisect import bisect_right
+from collections.abc import Callable
 from functools import partial, wraps
 from pathlib import Path
 from tempfile import NamedTemporaryFile
-from typing import Callable, Union
 
 import atlite
 import fiona
@@ -1099,7 +1099,7 @@ def rename_techs(label: str) -> str:
 
 
 def load_cutout(
-    cutout_files: Union[str, list[str]], time: Union[None, pd.DatetimeIndex] = None
+    cutout_files: str | list[str], time: None | pd.DatetimeIndex = None
 ) -> atlite.Cutout:
     """
     Load and optionally combine multiple cutout files.
@@ -1408,3 +1408,21 @@ def convert_units(
         )
 
     return df
+
+
+def check_cyear(cyear: int, scenario: str) -> int:
+    """Check if the climatic year is valid for the given scenario."""
+
+    valid_years = {
+        "NT": np.arange(1983, 2018).tolist(),
+        "DE": [1995, 2008, 2009],
+        "GA": [1995, 2008, 2009],
+    }
+
+    if cyear not in valid_years[scenario]:
+        logger.warning(
+            f"Snapshot year {cyear} doesn't match available TYNDP data. Falling back to 2009."
+        )
+        cyear = 2009
+
+    return cyear